+++ Angular2Emulator/TypeScript/Ng2Emulation/Templates/Parser.ts
@@ -1,5 +1,6 @@
 ﻿import {ParserRule} from "./ParserRule";
 import NgContentRule from "./rules/NgContentRule"
+import NgModelRule from "./rules/NgModelRule"
 import NgPropertyRule from "./rules/NgPropertyRule"
 import EventBindingRule from "./rules/EventBindingRule"
 import PropertyBindingRule from "./rules/PropertyBindingRule"
@@ -7,7 +8,7 @@ import TwoWayBindingRule from "./rules/TwoWayBindingRule"
 
 export default class Parser {
     private static rules: ParserRule[] = [
-		new NgContentRule(), new NgPropertyRule(),
+		new NgModelRule(), new NgContentRule(), new NgPropertyRule(),
 		new TwoWayBindingRule(), new PropertyBindingRule(), new EventBindingRule()
 	];